+++ sandbox/src/App.tsx
@@ -1,12 +1,13 @@
 // Sandbox to test the library. I tried using the symlink npx workaround but it's somewhat bugged.
 // Using a copy of the lib code here.
 import React, { useState } from 'react';
-import { StyleSheet, View, Text, Switch, TextInput, Pressable } from 'react-native';
-import { Shadow } from './index';
-import Slider from '@react-native-community/slider';
-import tinycolor from 'tinycolor2';
-import { PageScrollView } from 'pagescrollview';
+import { Pressable, StyleSheet, Switch, Text, TextInput, View } from 'react-native';
 import RadioForm from 'react-native-simple-radio-button';
+import { Slider } from '@sharcoux/slider';
+import { PageScrollView } from 'pagescrollview';
+import tinycolor from 'tinycolor2';
+import { Shadow } from './src/index'; // Aliased in Sandbox in dev.
+
 
 
 export const App: React.FC = () => {
@@ -15,7 +16,10 @@ export const App: React.FC = () => {
   const [offsetX, setOffsetX] = useState(0);
   const [offsetY, setOffsetY] = useState(0);
   const [paintInside, setPaintInside] = useState<boolean | undefined>(undefined);
+
   const [getChildRadius, setGetChildRadius] = useState(true);
+  const [getViewStyleRadius, setGetViewStyleRadius] = useState(true);
+
   const [size, setSize] = useState([defaults.width, defaults.height] as [number, number]);
   const [doUseSizeProp, setDoUseSizeProp] = useState(false);
 
@@ -25,6 +29,8 @@ export const App: React.FC = () => {
   const [startColor, setStartColor] = useState(defaults.startColor);
   const [finalColor, setFinalColor] = useState(defaults.finalColor);
   const [childColor, setChildColor] = useState(defaults.childColor);
+
+  // const [inset, setInset] = useState(true);
 
   return (
     <PageScrollView viewStyle={styles.container}>
@@ -38,10 +44,10 @@ export const App: React.FC = () => {
           <MySlider name='Child Width' step={0.1} range={[0, 200]} value={childWidth} onValueChange={setChildWidth}/>
           <MySlider name='Child Height' step={0.1} range={[0, 200]} value={childHeight} onValueChange={setChildHeight}/>
           <MySlider name='Distance' value={distance} onValueChange={setDistance}
-            range={[-10, 100]}  // min -10 to show < 0 won't do anything
+            range={[-10, 100]} // min -10 to show < 0 won't do anything
           />
           <MySlider name='Border Radius' value={borderRadius} onValueChange={setBorderRadius}
-            range={[-10, 100]}  // min -10 to show < 0 won't do anything
+            range={[-10, 100]} // min -10 to show < 0 won't do anything
           />
           <MySlider name='Offset X' range={[-20, 20]} value={offsetX} onValueChange={setOffsetX}/>
           <MySlider name='Offset Y' range={[-20, 20]} value={offsetY} onValueChange={setOffsetY}/>
@@ -78,9 +84,16 @@ export const App: React.FC = () => {
             labelHorizontal={false}
           />
 
+          {/* <MySwitch name='getChildRadius' value={getChildRadius} onValueChange={setGetChildRadius}/>
+          <MySwitch name='getViewStyleRadius' value={getViewStyleRadius} onValueChange={setGetViewStyleRadius}/> */}
+
+
           <MySwitch name='Use Size Prop' value={doUseSizeProp} onValueChange={setDoUseSizeProp}/>
           <MySlider name='Size Width Prop' step={0.1} range={[0, 200]} value={size[0]} onValueChange={v=>setSize([v, size[1]])}/>
           <MySlider name='Size Height Prop' step={0.1} range={[0, 200]} value={size[1]} onValueChange={v=>setSize([size[0], v])}/>
+
+          {/* <MySwitch name='Inset' value={inset} onValueChange={setInset}/> */}
+
 
         </View>
 
@@ -92,10 +105,14 @@ export const App: React.FC = () => {
             offset={(offsetX || offsetY) ? [offsetX, offsetY] : undefined} // To test paintInside default
             paintInside={paintInside}
             getChildRadiusStyle={getChildRadius}
-            radius={getChildRadius ? undefined : borderRadius}
+            getViewStyleRadius={getViewStyleRadius}
             containerViewStyle={{ margin: 100 }}
             size={doUseSizeProp ? size : undefined}
-            viewStyle={doUseSizeProp ? { backgroundColor: childColor } : undefined}
+            radius={getChildRadius ? undefined : borderRadius}
+            // TopEnd to check if it's supporting the Start/End combinations. When uncommenting this, also comment radius prop above.
+            // viewStyle={[doUseSizeProp && { backgroundColor: childColor }, { borderTopLeftRadius: 100, borderTopEndRadius: 10 }]}
+            viewStyle={[doUseSizeProp && { backgroundColor: childColor }]}
+
           >
             <View style={[
               !doUseSizeProp && { width: childWidth, height: childHeight },
@@ -108,7 +125,6 @@ export const App: React.FC = () => {
               },
             ]}/>
           </Shadow>
-          {/* <Text>aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa</Text> */}
         </View>
       </View>
 
@@ -125,12 +141,12 @@ const defaults = {
   height: 200,
   startColor: tinycolor('#00000020').toHex8String(),
   finalColor: tinycolor('#0000').toHex8String(),
-  childColor: tinycolor('#fff').toHex8String(),
+  childColor: tinycolor('#fff').toHex8String(), // tinycolor('#fff').toHex8String(),
 };
 
 
 const NameValue: React.FC<{
-  name: string, value: string | number | boolean | undefined, valueMonospace?: boolean
+  name: string; value: string | number | boolean | undefined; valueMonospace?: boolean;
 }> = ({ name, value, valueMonospace = false }) => {
   const prettyValue = typeof value === 'number' ? value.toFixed(1).replace(/[.,]0+$/, '') : String(value); // https://stackoverflow.com/a/5623195/10247962
   return (
@@ -148,12 +164,12 @@ const MySlider: React.FC<{
   name: string;
   step?: number;
   range: [min: number, max: number];
-  value: number
-  onValueChange: (value: number) => void
+  value: number;
+  onValueChange: (value: number) => void;
 }> = ({ name, step = 1, range, value, onValueChange }) => {
   return (
     <View style={{ marginBottom: 18 }}>
-      <NameValue {...{ name, value }} />
+      <NameValue {...{ name, value }}/>
       <View style={{ flexDirection: 'row' }}>
         <Pressable onPress={() => onValueChange(value - step)} style={({ pressed }) => [styles.decIncButton, pressed && { backgroundColor: '#bbb' }]}>
           <Text selectable={false} style={{ fontSize: 16, fontWeight: 'bold' }}>{'-'}</Text>
@@ -177,7 +193,7 @@ const MySwitch: React.FC<{
 }> = ({ name, onValueChange, value }) => {
   return (<>
     <NameValue name={name} value={value}/>
-    <Switch value={value} onValueChange={onValueChange} style={styles.switch} />
+    <Switch value={value} onValueChange={onValueChange} style={styles.switch}/>
   </>);
 };
 
@@ -241,7 +257,7 @@ const styles = StyleSheet.create({
   },
   decIncButton: {
     backgroundColor: '#fff',
-    padding: 8,
+    padding: 5,
     paddingHorizontal: 12,
     borderRadius: 4,
   },